Changed hash function to work on 32bit and activated correct test case.
The hash function used for AStar3D was, as Martin pointer out, not 32bit
ready. Changed that. Also activated a correct test case and commented out
the interim one.

+++ source/astar2.cpp
@@ -796,7 +796,7 @@ void AStar3D::findPath( const MapCoordinate3D& A, const vector<MapCoordinate3D>&
                   if ( !fa )
                      continue;
                   bool canStop = fa >= 2;
-                  DistanceType k = max(getMoveCost( N.h, hn, veh, hasAttacked ), 1);
+                  DistanceType k = max(getMoveCost( N.h, hn, veh, hasAttacked ), (float)1);
                   if (k >= longestPath)
                      continue;
